\name{yogurts}

\alias{yogurts}
\alias{yogurts-package}

\docType{package}

\title{yogurts!}

\description{
The yogurts package is a collection of ... algorithms.
}

\details{
\tabular{ll}{
Package: \tab yogurts\cr
Type: \tab Package\cr
Version: \tab 0.1.0\cr
Date: \tab 2018-03-14\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab Yes\cr
Depends: \tab Rcpp, RcppArmadillo \cr }
}
\section{How to cite this package}{
 \preformatted{@Manual{yogurts2018,
 author       = {Odran Bonnet, Alfred Galichon, Keith O'Hara, and Matt Shum},
 title        = {{yogurts}: Algorithms for ... using R and C++.},
 year         = {2018},
 note         = {R package version 0.1.0.}}}
\emph{}
}

\section{License}{
yogurts is licensed under the GNU General Public License (GPL) version 2, or (at your option) any later version.
}

\author{Odran Bonnet, Alfred Galichon, Keith O'Hara, and Matt Shum}